requirements.txt
shiny>=1.4.0
numpy>=1.24.0
matplotlib>=3.7.0
opencv-python>=4.8.0 
pandas >= 2.2.0
tensorflow>=2.12.0
scipy>=1.10.0
pillow>=10.0.0
rsconnect-python>=1.25.2